pyproject.toml
[project]
name = "backend-module-module"
version = "1.2.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "flask>=3.1.1",
    "flask-caching>=2.3.1",
    "flask-limiter>=3.12",
    "flask-marshmallow>=1.3.0",
    "flask-sqlalchemy>=3.1.1",
    "flask-swagger>=0.2.14",
    "flask-swagger-ui>=5.21.0",
    "marshmallow>=4.0.0",
    "marshmallow-sqlalchemy>=1.4.2",
    "mysql-connector-python>=9.3.0",
    "python-jose>=3.5.0",
    "requests>=2.32.3",
]
